Perform varimax rotation on a low rank matrix factorization

Perform varimax rotation on a low rank matrix factorization

# S3 method for svd_like
vsp(
  x,
  rank,
  ...,
  centerer = NULL,
  scaler = NULL,
  recenter = FALSE,
  rescale = TRUE,
  rownames = NULL,
  colnames = NULL
)

Arguments

x
TODO

rank
TODO

...
TODO

centerer
TODO

scaler
TODO

Examples

library(fastadi)

mf <- adaptive_impute(ml100k, rank = 20, max_iter = 5)

#> INFO [2020-11-21 00:42:34] Using svd initialization.
#> INFO [2020-11-21 00:42:34] Done initializing.
#> INFO [2020-11-21 00:42:34] Beginning AdaptiveImpute (max 5 iterations).
#> INFO [2020-11-21 00:42:34] Checking convergence every 1 iteration(s).
#> INFO [2020-11-21 00:42:35] Iter 1 complete. delta = 0.141364942080261, alpha = 211.680043817722
#> INFO [2020-11-21 00:42:36] Iter 2 complete. delta = 0.0337279590653107, alpha = 149.270605766213
#> INFO [2020-11-21 00:42:36] Iter 3 complete. delta = 0.0140181419142469, alpha = 117.618041284888
#> INFO [2020-11-21 00:42:37] Iter 4 complete. delta = 0.00741491366888598, alpha = 98.5325304350175
#> INFO [2020-11-21 00:42:37] Iter 5 complete. delta = 0.00450016707720721, alpha = 85.7878105197386

#> Warning: 
#> Reached maximum allowed iterations. Returning early.

fa <- vsp(mf)